Adding data port to AMC controller and python bindings.

+++ src/utilities.hpp
@@ -8,7 +8,7 @@ namespace dunedaq {
 namespace tdemodules {
 
 void append_big_uint16(std::vector<uint8_t>& buffer, uint16_t value);
-void append_big_uint32(std::vector<uint8_t>& buffer, uint16_t value);
+void append_big_uint32(std::vector<uint8_t>& buffer, uint32_t value);
 
 } // namespace tdemodules
 } // namespace dunedaq
